♻️ refactor: ReviewCard 컴포넌트 분리 및 구조화 (#48)
